requirements.txt
beautifulsoup4==4.10.0
requests==2.26.0
python-docx==0.8.11
openpyxl==3.0.9
reportlab==3.6.1
